Removed the filter to add 'identity' info on selective connections
Ir-respective of the end-point being a SERVER or CLIENT, it may be
useful for RI and upper layers to know the true identity of the sender
of the message. Therefore, removed the filter which was checking for
'SERVER' only mode.

Change-Id: I21cda0a645d79df6df3a3a4113cccbac4a7ae8d6
Signed-off-by: Sachin Agrawal <sachin.agrawal@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2949
Reviewed-by: Dmitriy Zhuravlev <d.zhuravlev@samsung.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
(cherry picked from commit e3cb235a806f9da339f4bb9ae07f5ac5cd8b25f3)
Reviewed-on: https://gerrit.iotivity.org/gerrit/3145