tinydtls: fix and activate android debug.c build
This patch activates the build of debug.c on android and also fixes one
build issue with this code in android.

in_port_t is not defined in <netinet/in.h> in the android NDK, use
"unsigned short int" instead because this is also used by sin_port in
the Android NDK.
There is already a bug for Android open to fix this problem in Android
here: https://code.google.com/p/android/issues/detail?id=80713

With this change the tinydtls debug.c file compiles for me for Android.

Change-Id: Id03a319768357ea1bef3f86ac2342e61e2b26b79
Signed-off-by: Hauke Mehrtens <hauke.mehrtens@lantiq.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/3195
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sachin Agrawal <sachin.agrawal@intel.com>