Corrected DB query of PDM
To check the existence of link between two devices, We should use query with 
‘and’ not ‘or’. 
Previously link  was checked with ‘or’ thats why if we try to create multiple 
links involving one same device and other different devices then it was showing 
link already exists. for example if we create link between A-B then if we try 
to create link between A-C then we are getting error that link already exists.

Change-Id: I408658af4f423a641bab077439435711ecac876e
Signed-off-by: Randeep Singh <randeep.s@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/3361
Reviewed-by: Woochul Shim <woochul.shim@samsung.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sachin Agrawal <sachin.agrawal@intel.com>
(cherry picked from commit 370c560720ef374672d6a91d57cb7c72311c2182)
Reviewed-on: https://gerrit.iotivity.org/gerrit/3425