Change occlient to cancel observe on every registration.
In the case when there are multiple servers, occlient sends out observer
requests to all of them. On reaching 15 notifications, only one of the
observation registrations were cancelled.
As the handle for cancellation was maintained in a global variable, none
of the others were cancelled and the client continued to get
notifications and this led to some failed tests.
Changed client behavior to cancel observe wit the handle present in the
response.

This can be cherry-picked to 1.0.0-dev

Change-Id: I51d28620b140dbb4f28baef7aa3be93276f6de5f
Signed-off-by: Mandeep Shetty <mandeep.shetty@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/3535
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>