Multiple EDR connection in Tizen
To support multiple EDR connection, the multicast data will be
added to an arraylist and as on when a device is discovered,
the data will be sent.  Once the discovery is finished, the data
will be removed from the arraylist.
Also the current implementation does discovery during initialization alone,
with this commit discovery will be done everytime when a multicast data
is sent.

Change-Id: I6bedb72ab716b828244d79f6f54bdaa4c8005b2f
Signed-off-by: vimala.v <vimala.v@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/3625
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>