Add exception handling logic in things configuration
When calling constructResourceObject, it probably throws an exception.
To avoid a sudden crash for it, we add exception handling to it.

Additionaly, in oicgroup file, there is a condition statement which
may lead wrong way from what we expect. Thus we remove it.

Change-Id: Ibc12cd261ca875738446ad3b10b199ae15497654
Signed-off-by: Jihun Ha <jihun.ha@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/3865
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Madan Lanka <lanka.madan@samsung.com>
(cherry picked from commit d22a600bab5001dbe770894de100853833a637ef)
Reviewed-on: https://gerrit.iotivity.org/gerrit/3877