tinydtls: use types from inittype.h, fix musl libc
The SHA2 code currently uses u_int32_t and similar types without
defining them. This type is not defined by the POSIX standard, but many
libc implementations have it as a non standard extension. In the musl
libc it is not implemented and I am getting a compile error when using
u_int32_t, I should use uint32_t, which is done when
SHA2_USE_INTTYPES_H is set.

This fixes build with musl libs used in OpenWrt.

Change-Id: I485435ddb8b1a2359caedd335ab54f91ca5e3f3e
Signed-off-by: Hauke Mehrtens <hauke.mehrtens@lantiq.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/3953
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sachin Agrawal <sachin.agrawal@intel.com>