connectivity: fix getnameinfo() call for musl libc
The musl libc and also the libc from net/freebsd are currently checking
if the size of the sockaddr given to getnameinfo() matches the size for
this family and it fails when giving sizeof(struct sockaddr_storage).
This patch fixes this problem by making the caller also provide the
size of the struct sockaddr_storage.

I am currently trying to get a fix for this problem into musl libc, but
IoTivity should still fix this.

This also fixes a wrong error message.

Fix for [IOT-834]

Change-Id: Ie9b89523b5ba20e536c03e3ad85c65bd5ff2ba53
Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Reviewed-on: https://gerrit.iotivity.org/gerrit/3959
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>
(cherry picked from commit 9db6f3452c76af909218e2b561467fcd208b413d)
Reviewed-on: https://gerrit.iotivity.org/gerrit/4439