Changed operator T enable_if to be more closely linked to the variant
The link between the variant and the enable_if needs to be strong!
Previously, there was no direct link, just one by 'coincidence'.  This
patch introduces a pair of type traits that bind the enable_if directly
to the variant.  Note that the type traits also work with ANY variadic
type!

Change-Id: Ie5f8fbc92b4f6800625c58fd73c08bedf147f122
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/4021
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Ossama Othman <ossama.othman@intel.com>
Reviewed-by: Habib Virji <habib.virji@samsung.com>