Optimization for EDR and LE adapter
In EDR and LE adapter, a request can't be differentiated as unicast or multicast.
Hence when a multicast request was received in EDR or LE adapter,
a error response was sent which could have been ignored.
Especially in Hetrogenous network, we received many responses with payload
NULL for a multicast request.  Hence as an optimization, we are blocking the
error response for EDR and LE adapters.

Change-Id: Ie3766115e8b89e21c760ec8232ebd74d73fc39c0
Signed-off-by: vimala.v <vimala.v@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/4029
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Abhishek Sharma <ce.abhishek@samsung.com>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>