Whitespace as delimiter for value in key-value pair of json file.
Updated file reading mechanism. The file now reads line by line and trim it
instead of using default delimiter.
Moreover, If a value(single value) in a key-value pair of json file contains
whitespace, the value is split to make an array of values for that key.
Added separate mothod for the task
Removed extra white space, and spelling mistakes.
Corrected coding stabndard.

Signed-off-by: fayaz.sc <fayaz.sc@samsung.com>
Change-Id: I318d4af93fc9a09348f3f836fc1a6dc68cf748fc
Reviewed-on: https://gerrit.iotivity.org/gerrit/4079
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Mushfiqul Islam <i.mushfiq@samsung.com>