packaging: Bump to 1.0.0
Note, I kept using 'linux' as os target,
because It seems to work enough on Tizen:2.4 and Tizen:3.

Since, iotivity is also supported in Tizen's upstream section,
as 'tizen' target, better use that one.

It can be helpful to have booth with different configuration.

As said earlier it tested it on Tizen:2.4:Mobile TM1 device.

Change-Id: Ibef1dc18ae6a8028bb4be5127e979ec46b795f9c
Signed-off-by: Philippe Coval <philippe.coval@osg.samsung.com>