/oic/res response for the RD
Earlier collection payload was added in discovery payload to handle RD response in
a separate way in form of tags and links. But with recent changes and discussion
baseURI is mvoed to discovery payload.

With this change there is one response for /oic/res with an additional field for RD.

Change-Id: Ia29cc64ffb48062d72ee4b9a9b59013480f13ff1
Signed-off-by: Habib Virji <habib.virji@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/4611
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>