Clean up documentation warnings reported by doxygen
Most of the documentation cleanup is due to overloaded member functions.

The @overload tag was used to remove the warnings. All of the parameteres
are documented for both functions with the @see tag added for each
overloaded function. This is a balance between not repeating documentation
yet still keeping the output clean. To have clean output some repetition
is required.

The @override tag was removed from the move constructors since @override is
not a keyword understood by doxygen.  The best I could tell the documentation
should have been the same for the move constructors as it was for the the
copy constructors.

note: @overload automatically adds "This is an overloaded member function,
provided for convenience. It differs from the above function only in what
argument(s) it accepts." to the output documentation.

Change-Id: Id167dc8d38bcca65d91a8bd3d6e43f9ca291aa79
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/4815
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>