Wait for Linux BLE event loop to start before stopping.
This addresses a corner case where the LE transport is stopped before
it has fully started.  In that case, the thread running the GLib event
loop never exited since the event loop itself was still running.  That
ultimately caused the CA layer termination to block indefinitely on a
pthread_join().

Change-Id: I521c7a66c6541cb14df7d4ba098899f303e50999
Signed-off-by: Ossama Othman <ossama.othman@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/4853
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>