EMPTY ACK/RESET message Handling by Routing Manager
In case of EMPTY message, according to CoAP spec it can have only 4 bytes containing
Version, Message type, Token Length as 0, Response result as CA_EMPTY and
Message ID.  However Route Option is required for routing the message via multihop.
In order to resolve this issue, we defined a MSGType in the RouteOption which is set
to 0 for normal messages while ACK for Empty acknowledgement and RST for Empty
Reset messages.
When a client sends Empty acknowledgement for a Response received, a POST message
carries the MSGType via multiple hops and the border gateway translates it to Empty ACK

Change-Id: I9550d1ad76f4b6a83a90ee2f06cef1899a3e7564
Signed-off-by: vimala.v <vimala.v@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/4649
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>