RE ServerBuilder always included "oic.if.baseline" interface.
If User don't setting "oic.if.baseline" interface.
but, RE ServerBuilder always support "oic.if.baseline" interface.

setDefaultInterface API(Modified comment of header) and Modified Initialize of Builder Constructor
If it is not called, the interface passed to the constructor is the default.

Added, Duplicate check the resource property of interface and resourcetype.

Change-Id: Ie25f1135cc7c99fbf5fe1b04f3121b9e2707be21
Signed-off-by: jaesick.shin <jaesick.shin@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/5529
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>