fixed the bug related to coap over tcp pdu creation.
The first nibble of the frame is used to indecate the length of
the options/payload. and if a value is 13, an 8-bit unsigned integer
follows the initial byte and indicates the length of options/payload
minus 13. even in case the length of options/payload is 13, an 8-bit
unsigned interger has to be used to indicate 0.

Change-Id: I580e126755ae4b3018fe7eb11eae34035b0efa6c
Signed-off-by: hyuna0213.jo <hyuna0213.jo@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/5901
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jee Hyeok Kim <jihyeok13.kim@samsung.com>
(cherry picked from commit 417ac656bad3f3c84f5cbbc8a4c88b5426a4fe1d)
Reviewed-on: https://gerrit.iotivity.org/gerrit/5999
Reviewed-by: Jon A. Cruz <jon@joncruz.org>