Parse integers and single-precision floats in incoming CBOR streams
Incoming single-precision data is automatically converted to
double-precision upon parsing. Integers are kept as is, but if the user
requests the data as floating point, we convert on the fly. That is
inefficient if the call is performed more than once, but the appication
shouldn't do that anyway to avoid memory allocations.

Fixes IOT-981.

Change-Id: I11a23ec8442c40bf9f7affff143e9edd0d17eec7
Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/6239
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Habib Virji <habib.virji@samsung.com>