Change name of encoding type
There is no name of encoding type as "oic.sec.encoding.bytestream" in the spec.
But we need a type to handling priv/pub data as byte array because our db is binary.
So we define new name of encoding type as "oic.sec.encoding.none" that means no need to encoding.
And we have asked to add this type into RAML spec.

-Patch 1 : Init
-Patch 2 : Modify encoding type from encodign.none to encoding.raw.

Change-Id: I7ad3d8facd302f97426cd42f139fbc11a7682638
Signed-off-by: Jongsung Lee <js126.lee@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/6329
Reviewed-by: dongik Lee <dongik.lee@samsung.com>
Reviewed-by: Chul Lee <chuls.lee@samsung.com>
Reviewed-by: Yonggoo Kang <ygace.kang@samsung.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Randeep Singh <randeep.s@samsung.com>