The TCP port number discovery based UDP multicast message
The TCP port number will be part of the discovery payload.
so the response to the UDP multicast message, should be an UDP unicast
reply containing the TCP port number if TCP adapter is activated.
After receiving the discovery response, the client can use TCP port
number to communicate with the server via CoAP/TCP.

- added tcp port value in policy of discovery payload
- the dynamic use of tcp server port in TCP adapter
- TCP adapter changes to support IPv6

Change-Id: Ie93361abec0287f0f600a791a58e41cf5b381892
Signed-off-by: hyuna0213.jo <hyuna0213.jo@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/7483
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jon@joncruz.org>