Cleaned up the unused-parameter warning
unused parameters are typically the result of 3 things
- they are from a call back and are not needed for that instance of
  the call back
- introduced to maintain backward compatability for older APIs
- variables only used in debug mode for debuging reasons

Many projects that try to build without warnings cast the unused
paramenter to a void type to prevent the warning.  The cast can
be added to the project using a #define.

The OC_UNUSED macro has been added to the octypes header file
to be used when in one of the three given scenarios.

In addition the macro has been used to prevent the unused-parameter
warning for several files.

Change-Id: Ib10967df4a135f0863aebbca41479d8e62d1a27b
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/7757
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jon@joncruz.org>