Clean up warning
This cleans up the warning capture of variable with non-automatic
storage duration

The warning is given when lambda functions are trying to capture
variables that are static or global. The easy fix is to not
capture the variables. They will still be avalible to the lamba
function without the variable capture.

Change-Id: I00890ddf5cf6932d1114ff4ca305905b4b877cc6
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/7761
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jon@joncruz.org>