Update to server .json and .db files to allow /p and /d access, to address JIRA issue 1077.
The default oic_svr_db_server_justworks.json and .dat files, along
with _randompin.json and .dat, do not contain ACL entries to allow
access to /p and /d.  This may confuse vendors who will most likely
wish to change this behavior.  IoTivity should (by default) allow
access, and allow vendors to change to not allow it if they wish.

Note that /oic/res is accessible in the current version prior to
this fix; it is just /oic/p and /oic/d which are not accessible
until new ACLs are provisioned.

Change-Id: I0ca424dc73db366a5c8ceca3236887d90543286e
Signed-off-by: Nathan Heldt-Sheller <nathan.heldt-sheller@intel.com>
Signed-off-by: Ashwini Kumar <k.ashwini@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/7797
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Mushfiqul Islam <i.mushfiq@samsung.com>
Reviewed-by: Randeep Singh <randeep.s@samsung.com>