change the condition for BLE connected callback
As BLE connected callback means communication on BLE is available for user,
the callback should be called on DescriptorWrite success callback which is the last step
for BLE communication.
However, since there is a case that gatt descriptor can not be found,
the callback needs to be called on ServicesDiscovered callback in this case.

Change-Id: I9d0daa233a1b605e70857d9fcc584f9224f53e03
Signed-off-by: Jaewook Jung <jw0213.jung@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/8025
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jon@joncruz.org>