first part of experimental IoTivity Lite checkin
IoTivity Light is a source code base for building a Lite IoTivity
server.  It is based on IoTivity 1.0.0, with updates to work with
1.0.1 wire protocol.  It is much smaller (by half), isn't
subject to memory fragmentation (since memory is allocated only
for transactions), and is highly configurable.

It has the same C and C++ APIs as IoTivity master, and it uses
the network adapters from the IoTivity master.  Work on it was
stopped before IoTivity security work was mature, so it contains
no security provisioning code.

This patch only contains the resource tree since attempting a full
push to Gerrit resulted in "internal error" from the Gerrit server.
Remaining to be pushed is a copy of the older version of tinycbor
it works with (along with minor patches to that tinycbor.)

This code is not expected to be merged into master because it
makes so many changes, but it provides many lessons (and data
structures) that may be useful for IoTivity Lite and IoTivity 2.0.

Change-Id: If7423504d9acfe4ad5abcb29f517c12c0febcc54
Signed-off-by: John Light <john.j.light@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/8669
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>