Clean compiler warnings.
in caipserver.c:
the function static void CAHanderlNetLink() was not defined or used and could
be removed. [-Wunused-function]

the function CAIPInitializeNetworkMonitorList was not returning a status on
the succesful use case. [-Wreturn-type]

In the CARemovedNetworkMonitorList the ifindex value comes from a system
library and can not be changed so I cast the removedifitem->index to int
to prevent the warning [-Wsign-compare]

Change-Id: I5dd8c52c2e82d68df9a0baf4ad2aa6acb4facfc7
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/8407
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: John Light <john.j.light@intel.com>
Reviewed-by: Jon A. Cruz <jon@joncruz.org>