[Update] process of provisioning was split.
Previously, Easy Setup for Mediator provided only 1 API, startProvisioning,
which was responsible for whole processes of easy setup.
We thought this approach had a limitation in case of user interaction required scenarios.
For this reason, we provide multiple APIs rather than a single API and
each of the processes of easy setup is as follows:

1. Discovery
2. SecurityProvisioning
3. Request Enrollee's resources
4. Data Provisioning (WiFi AP, device configuration information)
5. Cloud data provisioning (Auth Code, not implemented yet)

Change-Id: I172b9130531a0bb9e5b0f74aaee40a07523dcbd0
Signed-off-by: Heewon Park <h_w.park@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/8427
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>