Modified consumer internal function call flow.
When receive provider discover and message and syncinfo,
callback is called by communication module until this patch.
but, a proper way is callback should be calling by internalTaskControl module.
so, execution of callback logic is moved to internalTaskControl module.

Public NSProvider structure is modified.
Query Paramater(Consumer Id) is inclued on the request for subscription to providers.

Change-Id: Ib4e0db852d68889b90b66f880a052ed1453c2aca
Signed-off-by: KIM JungYong <jyong2.kim@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/8693
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>