Fixed policy response in /oic/res discovery
Added "sec": false to response in case of unsecure resource
Removed "port" information in case of unsecure resource

This patch is based on the Core Spec CR on resource policy.
If "sec" is false, it may or may not be included in /oic/res discovery response.
If "sec" is false, "port" shall be omitted in the response.
This patch should be used together with
Disable OC_SECURE flag in case of SECURED=0 build patch
(https://gerrit.iotivity.org/gerrit/#/c/8875/)

Change-Id: I8e0f3b39896c8856ab659f966e1227196e0f2153
Signed-off-by: Jongmin Choi <jminl.choi@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/8987
Reviewed-by: Habib Virji <habib.virji@samsung.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Randeep Singh <randeep.s@samsung.com>