[IOT-986] Not send response to multicast message for resource not existing
The resource not existing, it should not return any response for multicast packet.
There was no check for multicast request, this has been added and it does not send
any response.

Also add OC_MULTICAST type in octypes.h.
octypes.h transport flag should correspond to flags set in the CA layer. In the CA
layer (1 << 7), 7 bit is set for Multicast. In the RI layer, the type were differing,
hence was not possible to check if it is multicast packet. Please note the OC_RESERVED1
was not used anywhere and was wrong, it should match the transport flag of the CA layer.

BUG: http://jira.iotivity.org/browse/IOT-1126,
     http://jira.iotivity.org/browse/IOT-1107,
     http://jira.iotivity.org/browse/IOT-986

Change-Id: Ia8a7d1bdacf4e9a8a8d6cb21c49ad830cce7c3cc
Signed-off-by: Habib Virji <habib.virji@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/9073
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Mushfiqul Islam <i.mushfiq@samsung.com>
Reviewed-by: Ziran Sun <ziran.sun@samsung.com>