RD Payload memory handling
By default 255 bytes of memory is allocated for the payload, when it need
extra amount of memory, cbor library provides a size member that keeps incrementing
but no memory is allocated or written.

In case of failure this value is passed to the ocpayloadconvert.c, which allocates
the right size and calls the conversion function again. It is necessary to allow the
code to get the right size.

Change-Id: I92202a91780fcf01d4ff8fea0a3c194ac5a42e83
Signed-off-by: Habib Virji <habib.virji@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/9145
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>