Change request method to POST on updating cred in OTM
the issue is that the /cred resource will actually exist, just will have no creds in it (empty array).
So this is really an UPDATE, to update the "creds" property of the oic.r.cred resource.
Once CTT has changed to POST on OTM, OC_REST_PUT case will be removed in CredEntityHandler

Patch1: apply patch
Patch2: Add comment

Change-Id: Iaa4b7d72e5855575be1f91b48997ee445941df03
Signed-off-by: js126.lee <js126.lee@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/9161
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Nathan Heldt-Sheller <nathan.heldt-sheller@intel.com>
Reviewed-by: Randeep Singh <randeep.s@samsung.com>