Fix data alignment issue in tinydtls session_t
Detected a data alignment bug on Windows. The offset of the .addr union
is different in session_t vs stCADtlsAddrInfo_t.  With this fix,
SimpleServer should not crash when built in secured mode. Without this
fix, an exception will occur in dtls_security_params_epoch().

Verified structure accuracy on VS2015 using cdb.exe. Prior to this patch,
session_t.addr was at +4 and stCADtlsAddrInfo_t.addr was at +8.

Change-Id: I07439b19fe08474dbf0d46a315dfd5b7355c1bdc
Signed-off-by: David Antler <david.a.antler@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/9227
Reviewed-by: Jon A. Cruz <jon@joncruz.org>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>